requirements.txt
joblib==1.0.1
numpy==1.19.4
Pillow==8.0.1
scipy==1.7.0
torch==1.7.0
torchvision==0.8.1
numpy==1.19.4
PyYAML==6.0
